+++ extra/files/gen2-kali2021/zeroing.sh
@@ -1,0 +1,6 @@
+#!/usr/bin/env bash
+# Zeroing
+echo "Zeroing phase"
+  time dd if=/dev/zero|pv -treb|dd of=/file.zero bs=4096;sync;sync;rm -rfv /file.zero;sync;sync
+  rm -rfv /file.zero||true
+echo "Zeroing done"