+++ hv_rockylinux86.ps1
@@ -5,8 +5,8 @@ $startDTM = (Get-Date)
 
 # Variables
 $template_file="./templates/hv_rockylinux8_g2.pkr.hcl"
-$var_file="./variables/variables_rockylinux85.pkvars.hcl"
-$machine="RockyLinux 8.5"
+$var_file="./variables/variables_rockylinux86.pkvars.hcl"
+$machine="RockyLinux 8.6"
 $packer_log=0
 
 if ((Test-Path -Path "$template_file") -and (Test-Path -Path "$var_file")) {